refactor(useChatAI): extract assistant post-processing pipeline
Phase 0 of instant-push-agentic-loop plan. Pure refactor — local
fetch path behavior is byte-level identical. Extracts ~1500 lines
of pipeline (normalize / 2nd-pass LLM / actions / thinking / HTML /
sanitize / bilingual / chunk / persist) into utils/applyAssistantPostProcessing.ts
so the instant-push code path can call it in Phase 1.

ctx interface includes skipSecondPassLLM and directives parameters
that are no-ops in Phase 0; they enable Phase 1 and Phase 2 to
adjust behavior without changing the function signature.
